+++ edx_proctoring/api.py
@@ -11,7 +11,6 @@ from datetime import datetime, timedelta
 
 import pytz
 from opaque_keys import InvalidKeyError
-from opaque_keys.edx.keys import CourseKey, UsageKey
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -65,6 +64,7 @@ from edx_proctoring.utils import (
     humanized_time,
     is_reattempting_exam,
     obscured_user_id,
+    resolve_exam_url_for_learning_mfe,
     verify_and_add_wait_deadline
 )
 
@@ -107,6 +107,7 @@ def get_proctoring_settings_by_exam_id(exam_id):
             'provider_name': provider.verbose_name,
             'learner_notification_from_email': provider.learner_notification_from_email,
             'integration_specific_email': get_integration_specific_email(provider),
+            'proctoring_escalation_email': _get_proctoring_escalation_email(exam['course_id']),
         })
     return proctoring_settings_data
 
@@ -612,12 +613,16 @@ def get_exam_attempt_data(exam_id, attempt_id, is_learning_mfe=False):
         critically_low_threshold_pct * float(allowed_time_limit_mins) * 60
     )
 
+    if (not allowed_time_limit_mins
+            or (attempt and attempt['status'] == ProctoredExamStudentAttemptStatus.ready_to_resume)):
+        allowed_time_limit_mins = _calculate_allowed_mins(exam, attempt['user']['id'])
+
+    total_time = humanized_time(allowed_time_limit_mins)
+
     # resolve the LMS url, note we can't assume we're running in
     # a same process as the LMS
     if is_learning_mfe:
-        course_key = CourseKey.from_string(exam['course_id'])
-        usage_key = UsageKey.from_string(exam['content_id'])
-        exam_url_path = '{}/course/{}/{}'.format(settings.LEARNING_MICROFRONTEND_URL, course_key, usage_key)
+        exam_url_path = resolve_exam_url_for_learning_mfe(exam['course_id'], exam['content_id'])
 
     else:
         exam_url_path = reverse('jump_to', args=[exam['course_id'], exam['content_id']])
@@ -629,6 +634,7 @@ def get_exam_attempt_data(exam_id, attempt_id, is_learning_mfe=False):
         'exam_display_name': exam['exam_name'],
         'exam_url_path': exam_url_path,
         'time_remaining_seconds': time_remaining_seconds,
+        'total_time': total_time,
         'low_threshold_sec': low_threshold,
         'critically_low_threshold_sec': critically_low_threshold,
         'course_id': exam['course_id'],